move labelembedding to others
